Reorganized and reworked timeout handling
I'll be honest, I'm... not sure the old code worked?  I think maybe it
was using Contexts wrong.

Anyway, I refactored, rewrote, and simplified this code a fair bit by
creating a generic timing out reading function that takes a function
which does the reading.

I'll be honest, I'm 80% sure this is okay but man I gotta test it...